+++ db/migrate/20160114104102_add_mileage_to_vehicle.rb
@@ -1,5 +1,5 @@
 class AddMileageToVehicle < ActiveRecord::Migration
   def change
-    add_column :vehicles, :mileage, :integer
+    add_column :vehicles, :mileage, :decimal
   end
 end
