+++ db/migrate/20160114104845_add_average_mileage_to_vehicle.rb
@@ -1,5 +1,5 @@
 class AddAverageMileageToVehicle < ActiveRecord::Migration
   def change
-    add_column :vehicles, :average_mileage, :integer
+    add_column :vehicles, :average_mileage, :decimal
   end
 end
